refactor: 커스텀 예외 패키지 생성
