+++ jdbc/src/main/java/com/interface21/jdbc/exception/JdbcException.java
@@ -1,4 +1,4 @@
-package com.interface21.jdbc;
+package com.interface21.jdbc.exception;
 
 public class JdbcException extends RuntimeException {
