+++ jdbc/src/main/java/com/interface21/jdbc/exception/NoDataJdbcException.java
@@ -1,4 +1,4 @@
-package com.interface21.jdbc;
+package com.interface21.jdbc.exception;
 
 public class NoDataJdbcException extends JdbcException {
